pyproject.toml
[tool.poetry]
name = "api_sad_sig"
version = "0.1.0"
description = ""
authors = ["invarii <invaridewi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8,<3.9"
django = "^3.1.1"
djangorestframework = "^3.11.1"
djangorestframework-simplejwt = "^4.4.0"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"
pylint = "^2.6.0"
autopep8 = "^1.5.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
